src/main/java/com/example/demo/service/EmployeeService.java
package com.example.demo.service;

import java.util.List;

import org.springframework.stereotype.Service;

import com.example.demo.dao.EmployeeRepository;
import com.example.demo.entity.Employee;

@Service
public class EmployeeService {

    private final EmployeeRepository employeeRepository;

    public EmployeeService(EmployeeRepository employeeRepository) {
        this.employeeRepository = employeeRepository;
    }

    public Employee saveEmployee(Employee employee) {
    	return employeeRepository.save(employee);
    }

	public List<Employee> getAllEmployee() 
	{
		return employeeRepository.findAll();
	}
	
	
}
